src/requirements.txt
Django==4.2
djangorestframework==3.14.0
Pillow==9.5.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
gunicorn==20.1.0
psycopg2-binary==2.9.6
django-filter==23.2